+++ Cargo.toml
@@ -24,3 +24,5 @@ tracing-subscriber = "0.3"
 thiserror = "1.0"
 anyhow = "1.0"
 async-trait = "0.1"
+tokio = { version = "1.0", features = ["full"] }
+axum = "0.8"
