--- topkio-server/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "topkio-server"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-topkio-core = { path = "../topkio-core" }
-topkio-providers ={ path = "../topkio-providers" }
-axum = "0.6"
-tokio = { version = "1", features = ["full"] }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-anyhow = { workspace = true }
-
-[dev-dependencies]
-reqwest = { version = "0.11", features = ["json"] }
