--- topkio-server/src/metrix.rs
@@ -1,8 +1,0 @@
-// 占位：未来可添加 Prometheus 或其他监控逻辑
-pub struct Metrics;
-
-impl Metrics {
-    pub fn new() -> Self {
-        Metrics
-    }
-}
